requirements.txt
dj-database-url==2.3.0
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gunicorn==23.0.0
psycopg2-binary
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.10.0
